+++ src/Symfony/Component/Routing/Matcher/UrlMatcher.php
@@ -216,19 +216,8 @@ class UrlMatcher implements UrlMatcherInterface, RequestMatcherInterface
      *
      * @return array The first element represents the status, the second contains additional information
      */
-    protected function handleRouteRequirements(string $pathinfo, string $name, Route $route/* , array $routeParameters */): array
+    protected function handleRouteRequirements(string $pathinfo, string $name, Route $route, array $routeParameters): array
     {
-        if (\func_num_args() < 4) {
-            trigger_deprecation('symfony/routing', '6.1', 'The "%s()" method will have a new "array $routeParameters" argument in version 7.0, not defining it is deprecated.', __METHOD__);
-            $routeParameters = [];
-        } else {
-            $routeParameters = func_get_arg(3);
-
-            if (!\is_array($routeParameters)) {
-                throw new \TypeError(sprintf('"%s": Argument $routeParameters is expected to be an array, got "%s".', __METHOD__, get_debug_type($routeParameters)));
-            }
-        }
-
         // expression condition
         if ($route->getCondition() && !$this->getExpressionLanguage()->evaluate($route->getCondition(), [
             'context' => $this->context,