+++ src/Symfony/Component/Serializer/Normalizer/NormalizerInterface.php
@@ -18,8 +18,6 @@ use Symfony\Component\Serializer\Exception\LogicException;
 
 /**
  * @author Jordi Boggiano <j.boggiano@seld.be>
- *
- * @method getSupportedTypes(?string $format): array
  */
 interface NormalizerInterface
 {