+++ src/Symfony/Component/Filesystem/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+7.0
+---
+
+ * Add argument `$lock` to `Filesystem::appendToFile()`
+
 5.4
 ---